feat(devices): move deployment and maintenance log feedback from hooks to UI
Remove onSuccess/onError ReusableToast callbacks from useDeployDevice and
useAddMaintenanceLog hooks.
